+++ jkube-kit/resource/helm/src/test/java/org/eclipse/jkube/kit/resource/helm/HelmServiceTest.java
@@ -157,6 +157,7 @@ class HelmServiceTest {
       .keywords(Collections.singletonList("ci"))
       .maintainers(Collections.singletonList(Maintainer.builder().name("maintainer-from-config").build()))
       .icon("test-icon")
+      .appVersion("1.33.7")
       .engine("gotpl")
       .dependencies(Collections.singletonList(HelmDependency.builder().name("dependency-from-config").build()));
     // When
@@ -171,6 +172,7 @@ class HelmServiceTest {
       .hasFieldOrPropertyWithValue("description", "Description from helmconfig")
       .hasFieldOrPropertyWithValue("home", "https://example.com")
       .hasFieldOrPropertyWithValue("icon", "test-icon")
+      .hasFieldOrPropertyWithValue("appVersion", "1.33.7")
       .hasFieldOrPropertyWithValue("engine", "gotpl")
       .hasFieldOrPropertyWithValue("keywords", Collections.singletonList("fragment"))
       .hasFieldOrPropertyWithValue("sources", Collections.singletonList("https://source.example.com"))