+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Usage:
 * Fix #2390: support for all missing Chart.yaml fields
 * Fix #2444: Add support for Spring Boot application properties placeholders
 * Fix #2456: Add utility class to decompress archive files
+* Fix #2472: Support for Helm Chart.yaml appVersion field defaulting to project version
 
 ### 1.15.0 (2023-11-10)
 * Fix #2138: Support for Spring Boot Native Image